+++ mayan/apps/mime_types/backends/python_magic.py
@@ -11,12 +11,12 @@ class MIMETypeBackendPythonMagic(MIMETypeBackend):
     def _init(self, copy_length=None):
         self.copy_length = copy_length
 
-    def _get_mimetype(self, file_object, mimetype_only):
+    def _get_mime_type(self, file_object, mime_type_only):
         """
-        Determine a file's mimetype by calling the system's libmagic
+        Determine a file's MIME type by calling the system's libmagic
         library via python-magic.
         """
-        file_mimetype = None
+        file_mime_type = None
         file_mime_encoding = None
 
         with NamedTemporaryFile() as temporary_file_object:
@@ -30,18 +30,18 @@ class MIMETypeBackendPythonMagic(MIMETypeBackend):
 
             kwargs = {'mime': True}
 
-            if not mimetype_only:
+            if not mime_type_only:
                 kwargs['mime_encoding'] = True
 
             mime = magic.Magic(**kwargs)
 
-            if mimetype_only:
-                file_mimetype = mime.from_file(
+            if mime_type_only:
+                file_mime_type = mime.from_file(
                     filename=temporary_file_object.name
                 )
             else:
-                file_mimetype, file_mime_encoding = mime.from_file(
+                file_mime_type, file_mime_encoding = mime.from_file(
                     filename=temporary_file_object.name
                 ).split('; charset=')
 
-        return file_mimetype, file_mime_encoding
+        return file_mime_type, file_mime_encoding